Add a tox.ini.
Copied directly from the reporting pollster - not sure if that's
correct, but I don't see anything to suggest it won't work.

Also add a MANIFEST.in file.

The tox.ini file leaves out the hacking standard tests, as the current
code doesn't pass these and I'd like to avoid having a long list of
extraneous changes on top of the basic tox.ini fixes.

Change-Id: I78db9a14a77b322a75840122963418bf2e36adf4